Limit extruder speed
Tests with an e-step size of 0.001570796mm in reversed (unloading) showed
that the extruder stopped moving after switching from a feed rate of 2291
to 2292. If the calculations are correct then that would be ((micro)steps
per second)

24308 (2291 mm/min / 60 sec / 0.001570796 mm/e-step)
vs
24318 (2292 mm/min / 60 sec / 0.001570796 mm/e-step)

So the maximum is around 38mm/s. But after recalibrating the e-step size,
we found out that the actual orbiter (micro)step size is more like
0.0014492753623188406 mm/e-step.

24308 e-steps/s * 0.0014492753623188406 mm/e-step would be around
35.22898550724638 mm/s. So just round it down to be on the safe side.

Signed-off-by: Sven Eckelmann <sven@narfation.org>