Revert "Enable alternate wall printing directions"
While this option seems to work as expected and doesn't cause major
problems, it still seems to move start/stops of outer walls add unexpected
positions - which are too visible. So hidding a seam doesn't work as
expected.

So just disable it by default - but it might still be viable for various
models.

This reverts commit ec5dd19ea4ae75665ed3f237b0868cf632c18691.