xmax: Avoid scratches on top layer
If we directly move x+y to the homing position after the print stopped, it
can easily happn that the hot nozzle moves in a straight line over the just
ironed upper layer. This might cause "scratches" which can be seen on flat
surfaces.

Instead just move the nozzle 3mm away from the print and then home the
print head.

Signed-off-by: Sven Eckelmann <sven@narfation.org>