Set flow by default to 95% as starting point
It didn't turn out well to try to calibrate the flow by measuring two walls
with a caliber. Problem here is the underextrusion which will happen to get
the outer perimeter at the correct size.

The nozzle on a FDM printer is not able to print tangular, straight walls.
Instead, they are bulging outwards around of the middle (height) of each
line. The bulging was adding a couple of 0.0x mm on each line in the middle
but also reduced the top and bottom of each line by 0.0y mm. The low flow
multipliers were trying to reduce this effect in the middle by
under-extruding - but this also meant that the upper and bottom part of
each line were under-extruded.

This under-extrusion can cause instable prints because walls didn't
fuse/melded together - also the top layers might have gaps between lines.
This is also well explained in https://www.youtube.com/watch?v=xyrC0vVNGMs

Instead just use a middle ground (95%) to avoid overextrusion (due to
encapsulated air between walls of different layers) as recommened by the
prusa slicer. Fine tuning can be done later.

This might be moved later to the start/end code when it was tested
correctly.

Signed-off-by: Sven Eckelmann <sven@narfation.org>