xmax: Re-enable Z-Hop
It is recommended not to use Z-Hop on retract when linear/pressure advance is
enabled. But on thin models with a large height, it already happened that
some material on the nozzle collided with a printed part and ripped it of
the printbed.

To reduce the negative effects to linear/pressure advance, only use z-hop
when it needs to jump over printed parts and there is no other way around
it.

Signed-off-by: Sven Eckelmann <sven@narfation.org>