Add flow compensation for materials
The measurement was done at 0.2mm with two walls (each line 0.45mm wide).
The expected size of the wall was 0.9mm. The expected value was divided by
the (average of multiple) measured value to get an initial flow rate.

But to avoid too weak bonding between walls, a bulging of 0.02mm was used.
So the flow multiplier was multiplied by (92 / 90) to get to the values now
stored in the material files.

Signed-off-by: Sven Eckelmann <sven@narfation.org>