Only increase line width for first layer
We want roughly the same size as Prusa/Orcaslicer for our extrusions. But
they are using a completely different way to calculate the line spacings
and line widths (or actually the extrusion area). At the end, the
Prusa/Orcaslicer profile are printing with the same spacing of 0.4mm (when
precise walls are disabled). Just the first layer is printing slightly
higher to better connect it to the printbed.

Signed-off-by: Sven Eckelmann <sven@narfation.org>