Disable precise walls to keep them consistent
At the moment, we seem to have problems with overhangs during high speed
printing. This might partially be related to the not well connected outer
walls and has to be tested further without this setting enabled.

The actual precision might be better created by changing wall printing
order or fixing the flow compensation.

Signed-off-by: Sven Eckelmann <sven@narfation.org>